refactor backend for better file structure
