+++ napkin-backend/lib/responses/types.go
@@ -1,0 +1,7 @@
+package responses
+
+// HealthResponse represents the health check response
+type HealthResponse struct {
+	Status  string `json:"status"`
+	Message string `json:"message"`
+}
