+++ napkin-backend/main.go
@@ -1,112 +1,29 @@
 package main
 
 import (
-	"encoding/json"
 	"log"
 	"net/http"
-	"os"
-	"path/filepath"
-)
 
-type HealthResponse struct {
-	Status  string `json:"status"`
-	Message string `json:"message"`
-}
+	"napkin-backend/lib/handlers"
+	"napkin-backend/lib/middleware"
+	"napkin-backend/lib/static"
+)
 
 func main() {
 	mux := http.NewServeMux()
 
 	// API Routes
-	mux.HandleFunc("GET /health", healthHandler)
-	mux.HandleFunc("GET /api/hello", helloHandler)
+	mux.HandleFunc("GET /health", handlers.HealthHandler)
+	mux.HandleFunc("GET /api/hello", handlers.HelloHandler)
 
 	// Serve static files in production
-	staticDir := getStaticDir()
-	if staticDir != "" {
-		log.Printf("Serving static files from: %s", staticDir)
-		mux.Handle("/", spaHandler(staticDir))
-	}
+	static.SetupStaticHandler(mux)
 
 	// Enable CORS middleware (only needed in development when not using proxy)
-	handler := corsMiddleware(mux)
+	handler := middleware.CORS(mux)
 
 	log.Println("Starting server on :8080")
 	if err := http.ListenAndServe(":8080", handler); err != nil {
 		log.Fatal(err)
 	}
-}
-
-// getStaticDir returns the path to static files if they exist
-func getStaticDir() string {
-	// Check for dist folder relative to binary
-	candidates := []string{
-		"./dist",
-		"../napkin-app/dist",
-		"./napkin-app/dist",
-	}
-
-	for _, dir := range candidates {
-		if info, err := os.Stat(dir); err == nil && info.IsDir() {
-			absPath, _ := filepath.Abs(dir)
-			return absPath
-		}
-	}
-	return ""
-}
-
-// spaHandler serves static files and falls back to index.html for SPA routing
-func spaHandler(staticDir string) http.Handler {
-	fileServer := http.FileServer(http.Dir(staticDir))
-
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		path := filepath.Join(staticDir, r.URL.Path)
-
-		// Check if file exists
-		if info, err := os.Stat(path); err == nil && !info.IsDir() {
-			fileServer.ServeHTTP(w, r)
-			return
-		}
-
-		// Check if it's a directory with index.html
-		indexPath := filepath.Join(path, "index.html")
-		if _, err := os.Stat(indexPath); err == nil {
-			fileServer.ServeHTTP(w, r)
-			return
-		}
-
-		// Fallback to index.html for SPA client-side routing
-		http.ServeFile(w, r, filepath.Join(staticDir, "index.html"))
-	})
-}
-
-func corsMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Access-Control-Allow-Origin", "*")
-		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
-		w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
-
-		if r.Method == "OPTIONS" {
-			w.WriteHeader(http.StatusOK)
-			return
-		}
-
-		next.ServeHTTP(w, r)
-	})
-}
-
-func healthHandler(w http.ResponseWriter, r *http.Request) {
-	response := HealthResponse{
-		Status:  "ok",
-		Message: "Server is running",
-	}
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(response)
-}
-
-func helloHandler(w http.ResponseWriter, r *http.Request) {
-	response := map[string]string{
-		"message": "Hello from Napkin Backend!",
-	}
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(response)
 }
